+++ src/tractusx_sdk/dataspace/managers/oauth2_manager.py
@@ -1,6 +1,7 @@
 #################################################################################
 # Tractus-X - Industry Flag Service
 #
+# Copyright (c) 2025 LKS NEXT
 # Copyright (c) 2025 CGI Deutschland B.V. & Co. KG
 # Copyright (c) 2025 Contributors to the Eclipse Foundation
 #
@@ -22,8 +23,11 @@
 #################################################################################
 
 from keycloak.keycloak_openid import KeycloakOpenID
+from .auth_manager_interface import AuthManagerInterface
+from fastapi import Request, HTTPException
+from starlette.status import HTTP_401_UNAUTHORIZED
 
-class OAuth2Manager:
+class OAuth2Manager(AuthManagerInterface):
     
     """
     Class responsible for managing the IAM IDP Service
@@ -87,3 +91,23 @@ class OAuth2Manager:
         ## Build token header
         headers["Authorization"] = "Bearer " + self.get_token()
         return headers
+
+    def is_authenticated(self, request: Request) -> bool:
+        """
+        Check if the OAuth2Manager is authenticated.
+        This method verifies whether the Bearer token included in the request is valid.
+        :param request: FastAPI Request object.
+        :return: True if authenticated, False otherwise.
+        """
+        if not self.connected:
+            raise RuntimeError("Not connected. Please call the connect() method before checking authentication.")
+        
+        authorization: str = request.headers.get("Authorization")
+        if not authorization or not authorization.startswith("Bearer "):
+            raise HTTPException(status_code=HTTP_401_UNAUTHORIZED, detail="Missing or invalid Authorization header.")
+        token = authorization.split(" ")[1]
+        try:
+            user_info = self.keycloak_openid.userinfo(token)
+            return bool(user_info)
+        except Exception:
+            raise HTTPException(status_code=HTTP_401_UNAUTHORIZED, detail="Invalid or expired token.")