+++ src/tractusx_sdk/industry/__init__.py
@@ -21,6 +21,6 @@
 #################################################################################
 
 # Package-level variables
-__version__ = '0.2.0'
+__version__ = '0.3.2'
 __author__ = 'Eclipse Tractus-X Contributors'
 __license__ = "Apache License, Version 2.0"