+++ CHANGELOG.md
@@ -5,6 +5,25 @@ Further information can be found on the [README.md](README.md) file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.3.2] - 2025-07-22
+
+### Fixed
+
+- Fixed a bug in the memory connection manager and added missing logger support by @matbmoser in https://github.com/eclipse-tractusx/tractusx-sdk/pull/131
+
+## [0.3.1] - 2025-07-18 - not released, included in v0.3.2
+
+### Added
+
+- feat: enhance connection management with Postgres support + memory Postgres connection caching by @matbmoser in https://github.com/eclipse-tractusx/tractusx-sdk/pull/129
+
+## [0.3.0] - 2025-07-16
+
+- refactor(http-tools): update HttpTools methods to  avoid overriding by @samuelroywork in https://github.com/eclipse-tractusx/tractusx-sdk/pull/67
+- feat: added dependencies: Fixed conflicts in dependencies + session management by @matbmoser in https://github.com/eclipse-tractusx/tractusx-sdk/pull/124
+- feat: implement AuthManagerInterface and update authentication handling in managers by @CDiezRodriguez in https://github.com/eclipse-tractusx/tractusx-sdk/pull/122
+- feat: Simplify usage of SDK with better models + methods by @CDiezRodriguez in https://github.com/eclipse-tractusx/tractusx-sdk/pull/123
+
 ## [0.2.0] - 2025-07-14
 
 ### Added